improved constants modules
