--- cg/constants/application_info.py
@@ -1,8 +1,0 @@
-from cg.utils.enums import StrEnum
-
-
-class ApplicationInfo(StrEnum):
-    version: str = "version"
-    tag: str = "tag"
-    discounted_price: str = "discounted_price"
-    percent_kth: str = "percent_kth"
