--- cg/constants/costcenters.py
@@ -1,6 +1,0 @@
-from cg.utils.enums import StrEnum
-
-
-class Costcenters(StrEnum):
-    ki: str = "ki"
-    kth: str = "kth"
