--- cg/constants/customers.py
@@ -1,6 +1,0 @@
-from cg.utils.enums import StrEnum
-
-
-class CustomerNames(StrEnum):
-    cust999: str = "cust999"
-    cust032: str = "cust032"
