--- cg/constants/record_type.py
@@ -1,6 +1,0 @@
-from cg.utils.enums import StrEnum
-
-
-class RecordType(StrEnum):
-    Sample: str = "Sample"
-    Pool: str = "Pool"
